Changed double to BigDecimal

+++ app/src/main/java/com/example/chefgiraffe/domains/Item.java
@@ -1,13 +1,15 @@
 package com.example.chefgiraffe.domains;
 
+import java.math.BigDecimal;
+
 public class Item {
     private String id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String itemImageUrl;
 
-    public Item(String id, String name, String description, double price, String itemImageUrl) {
+    public Item(String id, String name, String description, BigDecimal price, String itemImageUrl) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -32,7 +34,7 @@ public class Item {
         return description;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
